# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VertexBuildMapModel": {
        "properties": {
          "vertex_builds": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/VertexBuildTable"
              },
              "type": "array"
            },
            "title": "Vertex Builds",
            "type": "object"
          }
        },
        "required": [
          "vertex_builds"
        ],
        "title": "VertexBuildMapModel",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "API key header": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "API key query": {
        "in": "query",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "OAuth2PasswordBearerCookie": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "HarxitFlow",
    "version": "1.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/monitor/builds": {
      "delete": {
        "operationId": "delete_vertex_builds_api_v1_monitor_builds_delete",
        "parameters": [
          {
            "in": "query",
            "name": "flow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Flow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "summary": "Delete Vertex Builds",
        "tags": [
          "Monitor"
        ]
      },
      "get": {
        "operationId": "get_vertex_builds_api_v1_monitor_builds_get",
        "parameters": [
          {
            "in": "query",
            "name": "flow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Flow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VertexBuildMapModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "summary": "Get Vertex Builds",
        "tags": [
          "Monitor"
        ]
      }
    },
    "/api/v1/monitor/messages": {
      "delete": {
        "operationId": "delete_messages_api_v1_monitor_messages_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "title": "Message Ids",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "summary": "Delete Messages",
        "tags": [
          "Monitor"
        ]
      },
      "get": {
        "operationId": "get_messages_api_v1_monitor_messages_get",
        "parameters": [
          {
            "in": "query",
            "name": "flow_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flow Id"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "sender",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sender"
            }
          },
          {
            "in": "query",
            "name": "sender_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sender Name"
            }
          },
          {
            "in": "query",
            "name": "order_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "timestamp",
              "title": "Order By"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get Messages Api V1 Monitor Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "summary": "Get Messages",
        "tags": [
          "Monitor"
        ]
      }
    },
    "/api/v1/monitor/messages/session/{old_session_id}": {
      "patch": {
        "operationId": "update_session_id_api_v1_monitor_messages_session__old_session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "old_session_id",
            "required": true,
            "schema": {
              "title": "Old Session Id",
              "type": "string"
            }
          },
          {
            "description": "The new session ID to update to",
            "in": "query",
            "name": "new_session_id",
            "required": true,
            "schema": {
              "description": "The new session ID to update to",
              "title": "New Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Update Session Id Api V1 Monitor Messages Session  Old Session Id  Patch",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "summary": "Update Session Id",
        "tags": [
          "Monitor"
        ]
      }
    },
    "/api/v1/monitor/messages/session/{session_id}": {
      "delete": {
        "description": "Delete messages for a single session.<br><br>Only deletes messages from sessions belonging to flows owned by the current user.",
        "operationId": "delete_messages_session_api_v1_monitor_messages_session__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "summary": "Delete Messages Session",
        "tags": [
          "Monitor"
        ]
      }
    },
...
    "/api/v1/monitor/messages/shared": {
      "get": {
        "description": "Get messages for a shared/public flow, scoped to the authenticated user.<br><br>Uses a deterministic virtual flow_id derived from the user's ID and the<br>original flow ID. Only messages stored under this virtual flow_id are returned.",
        "operationId": "get_shared_messages_api_v1_monitor_messages_shared_get",
        "parameters": [
          {
            "description": "The original public flow ID",
            "in": "query",
            "name": "source_flow_id",
            "required": true,
            "schema": {
              "description": "The original public flow ID",
              "format": "uuid",
              "title": "Source Flow Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "order_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "timestamp",
              "title": "Order By"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get Shared Messages Api V1 Monitor Messages Shared Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "summary": "Get Shared Messages",
        "tags": [
          "Monitor"
        ]
      }
    },
...
  }
}